[Feature] Add main blueprint
