+++ bucketlist/main/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+main = Blueprint('main', __name__)
+
+# Import last to prevent Import Error
+from . import views
+
