+++ bucketlist/views.py
@@ -23,34 +23,6 @@ def logged_in_user():
     return User.query.filter_by(username='wcyn').first()
 
 
-@app.route('/')
-@app.route('/index')
-def index():
-    data = {
-        "text": "hello there",
-        "color": "blue"
-    }
-    return jsonify(data)
-
-
-@app.errorhandler(404)
-def page_not_found(e):
-    data = {
-        "error": "404",
-        "message": "Page Not Found"
-    }
-    return jsonify(data), 404
-
-
-@app.errorhandler(500)
-def page_not_found(e):
-    data = {
-        "error": "500",
-        "message": "Internal Server Error"
-    }
-    return jsonify(data), 500
-
-
 @app.route('/add')
 def add():
     user = User(username="wcyn", email="wasonga.cynthia@gmail.com")
